+++ playwright_recaptcha/recaptchav3/sync_solver.py
@@ -17,11 +17,11 @@ class SyncSolver:
     ----------
     page : Page
         The Playwright page to solve the reCAPTCHA on.
-    timeout : int, optional
+    timeout : float, optional
         The solve timeout in seconds, by default 30.
     """
 
-    def __init__(self, page: Page, timeout: int = 30) -> None:
+    def __init__(self, page: Page, timeout: float = 30) -> None:
         self._page = page
         self._timeout = timeout
 
@@ -61,13 +61,13 @@ class SyncSolver:
         except KeyError:
             pass
 
-    def solve_recaptcha(self, timeout: Optional[int] = None) -> str:
+    def solve_recaptcha(self, timeout: Optional[float] = None) -> str:
         """
         Wait for the reCAPTCHA to be solved and return the `g-recaptcha-response` token.
 
         Parameters
         ----------
-        timeout : Optional[int], optional
+        timeout : Optional[float], optional
             The solve timeout in seconds, by default 30.
 
         Returns
